+++ next.config.js
@@ -1,18 +1,5 @@
-const withPWA = require('next-pwa');
-
-module.exports = withPWA({
-  pwa: {
-    dest: 'public',
-    register: true,
-    skipWaiting: true,
-    disable: process.env.NODE_ENV === 'development',
-  },
-  compiler: {
-    styledComponents: {
-      ssr: true,
-      cssProp: true,
-    },
-  },
+/** @type {import('next').NextConfig} */
+const nextConfig = {
   i18n: {
     locales: ['en'],
     defaultLocale: 'en',
@@ -29,4 +16,6 @@ module.exports = withPWA({
       },
     ],
   },
-});
+};
+
+module.exports = nextConfig;